設定ファイルを更新

+++ git-add.sh
@@ -18,9 +18,11 @@
 			mkdir -p chunk
 			cd chunk
 			(cat region_checksum.txt; cat "$tmp1") | sort | uniq -u | awk '{printf "../region/%s\n", $1}' > "$tmp2"
-			cat "$tmp2" | xargs "$split_regions_to_chunks"
-			cp "$tmp1" ./region_checksum.txt
-			git add region_checksum.txt
+			if [ -s "$tmp2" ]; then
+				cat "$tmp2" | xargs "$split_regions_to_chunks"
+				cp "$tmp1" ./region_checksum.txt
+				git add region_checksum.txt
+			fi
 		)
 	done
 	git add -A world world_nether world_the_end
